+++ Engine/Goal.h
@@ -6,21 +6,21 @@
 class Goal
 {
 public:
-	Goal( int in_x,int in_y )
+	Goal( float in_x,float in_y )
 		:
 		x( in_x ),
 		y( in_y )
 	{}
 	void Draw( Graphics& gfx ) const
 	{
-		gfx.DrawRectDim( x,y,dimension,dimension,c );
+		gfx.DrawRectDim( int( x ),int( y ),int( dimension ),int( dimension ),c );
 	}
 	bool TestCollision( const Dude& dude ) const
 	{
-		const int duderight = dude.GetX() + dude.GetWidth();
-		const int dudebottom = dude.GetY() + dude.GetHeight();
-		const int pooright = x + dimension;
-		const int poobottom = y + dimension;
+		const float duderight = dude.GetX() + dude.GetWidth();
+		const float dudebottom = dude.GetY() + dude.GetHeight();
+		const float pooright = x + dimension;
+		const float poobottom = y + dimension;
 
 		return
 			duderight >= x &&
@@ -28,7 +28,7 @@ public:
 			dudebottom >= y &&
 			dude.GetY() <= poobottom;
 	}
-	void Respawn( int in_x,int in_y )
+	void Respawn( float in_x,float in_y )
 	{
 		x = in_x;
 		y = in_y;
@@ -59,9 +59,9 @@ public:
 		}
 	}
 private:
-	static constexpr int dimension = 20;
+	static constexpr float dimension = 20;
 	Color c = { 127,0,0 };
 	bool colorIncreasing = true;
-	int x;
-	int y;
+	float x;
+	float y;
 };
